pipeline/requirements.txt
autogluon.tabular[all]
autoviz
black[jupyter]
boto3
cleanlab
dask[dataframe]
datasets
dtale
dvc[all]
dvclive
feast[aws]
flake8
hydra_colorlog
hydra-core
isort
joblib
matplotlib
missingno
numba<=0.58.1
numpy
omegaconf
pandas
pyjanitor
python-dotenv
python-pptx
ray
ruff
scikit-learn
seaborn
shap
sweetviz
tenacity
tensorflow
tensorflow-addons
tensorflow-cpu
ucimlrepo
virtualenv
ydata-profiling